New set of RPC fixes (#1498)
- Removed deadlocks caused in some cases by mutexes
- Enable oAuth for RPC emergency mode,  should work since policies are here
- Simplified `listen` reload logic
- Simplified keep alive function
- Restrict concurrent "Connect" calls
- Set default connection pool to 20 (Our hybrid docker image already have it but in config file)
- Fixed concurrency issue causing multiple reloads when backup is loaded
- Added tests for protected APIs when in backup mode

+++ rpc_storage_handler.go
@@ -65,28 +65,20 @@ func rpcKeepAliveCheck(r *RPCStorageHandler) {
 		return
 	}
 
-	// Make sure the auth back end is still alive
-	c1 := make(chan string, 1)
-
-	go func() {
-		log.Debug("Getting keyspace check test key")
-		r.GetKey("0000")
-		log.Debug("--> done")
-		c1 <- "1"
-		close(c1)
-	}()
-
-	select {
-	case res := <-c1:
-		log.Debug("RPC Still alive: ", res)
-		// Don't run too quickly, pulse every 10 secs
-		time.Sleep(time.Second * 10)
-	case <-time.After(time.Second * 10):
+	log.Debug("Getting keyspace check test key")
+	err := r.SetKey("0000", "0000", 10)
+	// This error message comes from RPC layer
+	if err != nil && err.Error() != "Write dissallowed for API Tokens" {
 		log.WithFields(logrus.Fields{
 			"prefix": "RPC Conn Mgr",
 		}).Warning("Handler seems to have disconnected, attempting reconnect")
-		r.ReConnect()
+	} else {
+		if err != nil {
+			log.Debug("RPC Still alive")
+		}
 	}
+
+	time.Sleep(time.Second * 10)
 }
 
 // RPCStorageHandler is a storage manager that uses the redis database.
@@ -151,11 +143,20 @@ func emitRPCErrorEventKv(jobName string, funcName string, err error, kv map[stri
 	}
 }
 
+var rpcConnectMu sync.Mutex
+
 // Connect will establish a connection to the DB
 func (r *RPCStorageHandler) Connect() bool {
+	rpcConnectMu.Lock()
+	defer rpcConnectMu.Unlock()
+
 	if RPCClientIsConnected {
 		log.Debug("Using RPC singleton for connection")
 		return true
+	}
+
+	if RPCCLientSingleton != nil {
+		return rpcEmergencyMode != true
 	}
 
 	// RPC Client is unset
@@ -187,21 +188,20 @@ func (r *RPCStorageHandler) Connect() bool {
 
 	RPCCLientSingleton.Conns = config.Global.SlaveOptions.RPCPoolSize
 	if config.Global.SlaveOptions.RPCPoolSize == 0 {
-		RPCCLientSingleton.Conns = 50
+		RPCCLientSingleton.Conns = 20
 	}
 
 	RPCCLientSingleton.Dial = func(addr string) (conn io.ReadWriteCloser, err error) {
-		configMu.Lock()
-		defer configMu.Unlock()
-
 		dialer := &net.Dialer{
 			Timeout:   10 * time.Second,
 			KeepAlive: 30 * time.Second,
 		}
 
-		useSSL := false
-		if config.Global.SlaveOptions.UseSSL {
-			useSSL = true
+		configMu.Lock()
+		useSSL := config.Global.SlaveOptions.UseSSL
+		configMu.Unlock()
+
+		if useSSL {
 			cfg := &tls.Config{
 				InsecureSkipVerify: config.Global.SlaveOptions.SSLInsecureSkipVerify,
 			}
@@ -282,15 +282,19 @@ func (r *RPCStorageHandler) cleanKey(keyName string) string {
 	return setKeyName
 }
 
+var rpcLoginMu sync.Mutex
+
 func (r *RPCStorageHandler) ReAttemptLogin(err error) bool {
+	rpcLoginMu.Lock()
 	log.Warning("[RPC Store] Login failed, waiting 3s to re-attempt")
 
 	if rpcLoadCount == 0 && !rpcEmergencyModeLoaded {
 		log.Warning("[RPC Store] --> Detected cold start, attempting to load from cache")
 		log.Warning("[RPC Store] ----> Found APIs... beginning emergency load")
-		doReload()
 		rpcEmergencyModeLoaded = true
+		go doReload()
 	}
+	rpcLoginMu.Unlock()
 
 	time.Sleep(time.Second * 3)
 	if strings.Contains(err.Error(), "Cannot obtain response during timeout") {
@@ -332,11 +336,11 @@ func (r *RPCStorageHandler) GroupLogin() bool {
 
 	// Recovery
 	if rpcEmergencyMode {
-		doReload()
+		rpcEmergencyMode = false
+		rpcEmergencyModeLoaded = false
+		reloadURLStructure(nil)
 	}
 
-	rpcEmergencyMode = false
-	rpcEmergencyModeLoaded = false
 	return true
 }
 
@@ -371,11 +375,11 @@ func (r *RPCStorageHandler) Login() bool {
 	rpcLoadCount++
 
 	if rpcEmergencyMode {
-		doReload()
+		rpcEmergencyMode = false
+		rpcEmergencyModeLoaded = false
+		reloadURLStructure(nil)
 	}
 
-	rpcEmergencyMode = false
-	rpcEmergencyModeLoaded = false
 	return true
 }
 
@@ -411,8 +415,9 @@ func (r *RPCStorageHandler) GetKey(keyName string) (string, error) {
 			},
 		)
 		if r.IsAccessError(err) {
-			r.Login()
-			return r.GetKey(keyName)
+			if r.Login() {
+				return r.GetKey(keyName)
+			}
 		}
 
 		log.Debug("Error trying to get value:", err)
@@ -448,8 +453,9 @@ func (r *RPCStorageHandler) GetExp(keyName string) (int64, error) {
 			},
 		)
 		if r.IsAccessError(err) {
-			r.Login()
-			return r.GetExp(keyName)
+			if r.Login() {
+				return r.GetExp(keyName)
+			}
 		}
 		log.Error("Error trying to get TTL: ", err)
 		return 0, storage.ErrKeyNotFound
@@ -482,10 +488,15 @@ func (r *RPCStorageHandler) SetKey(keyName, session string, timeout int64) error
 				"fixedKeyName": ibd.KeyName,
 			},
 		)
-	}
-	if r.IsAccessError(err) {
-		r.Login()
-		return r.SetKey(keyName, session, timeout)
+
+		if r.IsAccessError(err) {
+			if r.Login() {
+				return r.SetKey(keyName, session, timeout)
+			}
+		}
+
+		log.Debug("Error trying to set value:", err)
+		return err
 	}
 
 	elapsed := time.Since(start)
@@ -513,8 +524,10 @@ func (r *RPCStorageHandler) Decrement(keyName string) {
 		)
 	}
 	if r.IsAccessError(err) {
-		r.Login()
-		r.Decrement(keyName)
+		if r.Login() {
+			r.Decrement(keyName)
+			return
+		}
 	}
 }
 
@@ -538,8 +551,9 @@ func (r *RPCStorageHandler) IncrememntWithExpire(keyName string, expire int64) i
 		)
 	}
 	if r.IsAccessError(err) {
-		r.Login()
-		return r.IncrememntWithExpire(keyName, expire)
+		if r.Login() {
+			return r.IncrememntWithExpire(keyName, expire)
+		}
 	}
 
 	if val == nil {
@@ -573,10 +587,14 @@ func (r *RPCStorageHandler) GetKeysAndValuesWithFilter(filter string) map[string
 				"searchStr": searchStr,
 			},
 		)
-	}
-	if r.IsAccessError(err) {
-		r.Login()
-		return r.GetKeysAndValuesWithFilter(filter)
+
+		if r.IsAccessError(err) {
+			if r.Login() {
+				return r.GetKeysAndValuesWithFilter(filter)
+			}
+		}
+
+		return nil
 	}
 
 	returnValues := make(map[string]string)
@@ -596,10 +614,14 @@ func (r *RPCStorageHandler) GetKeysAndValues() map[string]string {
 	kvPair, err := RPCFuncClientSingleton.CallTimeout("GetKeysAndValues", searchStr, GlobalRPCCallTimeout)
 	if err != nil {
 		emitRPCErrorEvent(rpcFuncClientSingletonCall, "GetKeysAndValues", err)
-	}
-	if r.IsAccessError(err) {
-		r.Login()
-		return r.GetKeysAndValues()
+
+		if r.IsAccessError(err) {
+			if r.Login() {
+				return r.GetKeysAndValues()
+			}
+		}
+
+		return nil
 	}
 
 	returnValues := make(map[string]string)
@@ -627,10 +649,12 @@ func (r *RPCStorageHandler) DeleteKey(keyName string) bool {
 				"fixedKeyName": r.fixKey(keyName),
 			},
 		)
-	}
-	if r.IsAccessError(err) {
-		r.Login()
-		return r.DeleteKey(keyName)
+
+		if r.IsAccessError(err) {
+			if r.Login() {
+				return r.DeleteKey(keyName)
+			}
+		}
 	}
 
 	return ok == true
@@ -648,10 +672,12 @@ func (r *RPCStorageHandler) DeleteRawKey(keyName string) bool {
 				"keyName": keyName,
 			},
 		)
-	}
-	if r.IsAccessError(err) {
-		r.Login()
-		return r.DeleteRawKey(keyName)
+
+		if r.IsAccessError(err) {
+			if r.Login() {
+				return r.DeleteRawKey(keyName)
+			}
+		}
 	}
 
 	return ok == true
@@ -677,10 +703,12 @@ func (r *RPCStorageHandler) DeleteKeys(keys []string) bool {
 					"asInterface": strings.Join(asInterface, ","),
 				},
 			)
-		}
-		if r.IsAccessError(err) {
-			r.Login()
-			return r.DeleteKeys(keys)
+
+			if r.IsAccessError(err) {
+				if r.Login() {
+					return r.DeleteKeys(keys)
+				}
+			}
 		}
 
 		return ok == true
@@ -723,8 +751,9 @@ func (r *RPCStorageHandler) AppendToSet(keyName, value string) {
 		)
 	}
 	if r.IsAccessError(err) {
-		r.Login()
-		r.AppendToSet(keyName, value)
+		if r.Login() {
+			r.AppendToSet(keyName, value)
+		}
 	}
 }
 
@@ -748,10 +777,12 @@ func (r *RPCStorageHandler) SetRollingWindow(keyName string, per int64, val stri
 				"per":     strconv.Itoa(int(per)),
 			},
 		)
-	}
-	if r.IsAccessError(err) {
-		r.Login()
-		return r.SetRollingWindow(keyName, per, val, false)
+
+		if r.IsAccessError(err) {
+			if r.Login() {
+				return r.SetRollingWindow(keyName, per, val, false)
+			}
+		}
 	}
 
 	elapsed := time.Since(start)
@@ -804,10 +835,14 @@ func (r *RPCStorageHandler) GetApiDefinitions(orgId string, tags []string) strin
 				"tags":  strings.Join(tags, ","),
 			},
 		)
-	}
-	if r.IsAccessError(err) {
-		r.Login()
-		return r.GetApiDefinitions(orgId, tags)
+
+		if r.IsAccessError(err) {
+			if r.Login() {
+				return r.GetApiDefinitions(orgId, tags)
+			}
+		}
+
+		return ""
 	}
 	log.Debug("API Definitions retrieved")
 
@@ -830,10 +865,14 @@ func (r *RPCStorageHandler) GetPolicies(orgId string) string {
 				"orgId": orgId,
 			},
 		)
-	}
-	if r.IsAccessError(err) {
-		r.Login()
-		return r.GetPolicies(orgId)
+
+		if r.IsAccessError(err) {
+			if r.Login() {
+				return r.GetPolicies(orgId)
+			}
+		}
+
+		return ""
 	}
 
 	if defString != nil {
@@ -917,10 +956,12 @@ func (r *RPCStorageHandler) CheckForKeyspaceChanges(orgId string) {
 			reqData,
 		)
 		if r.IsAccessError(err) {
-			r.Login()
-			r.CheckForKeyspaceChanges(orgId)
+			if r.Login() {
+				r.CheckForKeyspaceChanges(orgId)
+			}
 		}
 		log.Warning("Keysapce warning: ", err)
+		return
 	}
 
 	if keys == nil {
